+++ resources/release-notes.md
@@ -1,3 +1,4 @@
 ## Software
 
 - (Bartington, HelioSwarm, IMAP) Add support for more date formats for metadata readers
+- (IMAP) Add sensor activation attempts and activation status to SID15 (`mag.imap.hk.SID15`)